Directory structure improved
