+++ include/gdata/vector.h
