+++ src/stack.c
@@ -1,4 +1,4 @@
-#include "stack.h"
+#include "gdata/stack.h"
 #include <stdlib.h>
 #include <string.h>
