+++ src/vector.c
@@ -1,4 +1,4 @@
-#include "vector.h"
+#include "gdata/vector.h"
 #include <stdlib.h>
 #include <string.h>
